Explicitly casting tag and sender ranks to uint32_t for version compatibilty

+++ cpp/comms/std/src/ucp_helper.h
@@ -45,7 +45,7 @@ static void recv_handle(void *request, ucs_status_t status,
 struct ucx_context *ucp_isend(ucp_ep_h ep_ptr, const void *buf, int size,
                               int tag, int rank) {
 
-  ucp_tag_t ucp_tag = (rank << 31) | tag;
+  ucp_tag_t ucp_tag = ((uint32_t)rank << 31) | (uint32_t)tag;
 
   struct ucx_context *ucp_request = (struct ucx_context *)ucp_tag_send_nb(
     ep_ptr, buf, size, ucp_dt_make_contig(1), ucp_tag, send_handle);
@@ -77,7 +77,7 @@ struct ucx_context *ucp_isend(ucp_ep_h ep_ptr, const void *buf, int size,
  */
 struct ucx_context *ucp_irecv(ucp_worker_h worker, ucp_ep_h ep_ptr, void *buf,
                               int size, int tag, int sender_rank) {
-  ucp_tag_t ucp_tag = (sender_rank << 31) | tag;
+  ucp_tag_t ucp_tag = ((uint32_t)sender_rank << 31) | (uint32_t)tag;
 
   struct ucx_context *ucp_request = (struct ucx_context *)ucp_tag_recv_nb(
     worker, buf, size, ucp_dt_make_contig(1), ucp_tag, default_tag_mask,
